pyproject.toml
[build-system]
requires = ["setuptools>=61.0,<70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentos-ai"
version = "1.1.7"
description = "Production-Ready AI Agent Runtime - Automate everything with LLM-powered agents"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "AgentOS Team", email = "support@agentos.dev"}
]
maintainers = [
    {name = "AgentOS Team", email = "support@agentos.dev"}
]
keywords = [
    "ai",
    "agent",
    "llm",
    "automation",
    "orchestration",
    "chatbot",
    "openai",
    "claude",
    "gemini",
    "ollama",
    "gpt",
    "artificial-intelligence",
    "machine-learning",
    "command-line",
    "cli",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]

# Core dependencies only - keep minimal for easy installation
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "ollama>=0.1.0",
    "flask>=2.3.0",
    "pywebview>=4.4.0",
    "PyQt5>=5.15.0",
    "PyQtWebEngine>=5.15.0",
    "QtPy>=2.0.0",
]

[project.optional-dependencies]
# LLM providers
openai = ["openai>=1.0.0"]
claude = ["anthropic>=0.18.0"]
gemini = ["google-generativeai>=0.3.0"]
cohere = ["cohere>=4.0.0"]
ollama = ["ollama>=0.1.0"]
mcp = ["mcp>=0.1.0"]

# All LLM providers
llm = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "cohere>=4.0.0",
    "ollama>=0.1.0",
]

# Web UI
web = [
    "flask>=2.3.0",
    "gunicorn>=21.0.0",
    "gevent>=23.0.0",
]

# Desktop UI
desktop = [
    "pywebview>=4.4.0",
    "PyQt5>=5.15.0",
    "PyQtWebEngine>=5.15.0",
    "QtPy>=2.0.0",
]

# Docker isolation
docker = ["docker>=6.0.0"]

# Monitoring
monitoring = ["prometheus-client>=0.19.0"]

# Development tools
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

# Full installation with everything
full = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    "cohere>=4.0.0",
    "ollama>=0.1.0",
    "mcp>=0.1.0",
    "flask>=2.3.0",
    "gunicorn>=21.0.0",
    "gevent>=23.0.0",
    "pywebview>=4.4.0",
    "docker>=6.0.0",
    "prometheus-client>=0.19.0",
]

[project.scripts]
agentos = "agentos.agentos:main"

[project.urls]
Homepage = "https://agentos.sodeom.com/"
Documentation = "https://agentos.sodeom.com/docs"
Repository = "https://github.com/agents-os/agentos"
Issues = "https://github.com/agents-os/agentos/issues"
Changelog = "https://github.com/agents-os/agentos/blob/main/MD/CHANGELOG.md"

[tool.setuptools]
packages = ["agentos", "agentos.agent", "agentos.cli", "agentos.core", "agentos.database", "agentos.llm", "agentos.mcp", "agentos.web"]
include-package-data = true

[tool.setuptools.package-data]
agentos = [
    "py.typed",
    "templates/*.html",
    "static/*.css",
    "static/*.js",
    "static/*.png",
    "static/*.ico",
]
"*" = ["*.yaml", "*.yml", "*.json", "*.md"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | build
    | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist", ".venv"]
